addProgressBarToTimelineActivity
