+++ app/src/main/res/layout/login.xml
@@ -41,14 +41,6 @@
         android:text="@string/delete_Accounts"
         android:onClick="deleteAccounts"/>
 
-    <ProgressBar
-        android:id="@+id/progressBar"
-        style="?android:attr/progressBarStyle"
-        android:layout_centerHorizontal="true"
-        android:layout_centerVertical="true"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:visibility="invisible" />
 
 
 </LinearLayout>
