+++ app/src/main/res/layout/main.xml
@@ -4,6 +4,12 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
+    <ProgressBar
+        android:id="@+id/timeline_loading_progress"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:visibility="visible" />
+
     <ListView
         android:id="@+id/list"
         android:layout_width="match_parent"
